Rename ValueHandler to MySQLDataTypeHandler (#16685)
* Refactor ValueHandler to impl StatelessTypedSPI

* Rename ValueHandler to MySQLDataTypeHandler
