+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-dialect/shardingsphere-data-pipeline-mysql/src/main/java/org/apache/shardingsphere/data/pipeline/mysql/ingest/column/value/MySQLDataTypeHandler.java
@@ -17,22 +17,20 @@
 
 package org.apache.shardingsphere.data.pipeline.mysql.ingest.column.value;
 
-import org.junit.Test;
+import org.apache.shardingsphere.spi.type.typed.StatelessTypedSPI;
 
 import java.io.Serializable;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
-
-public final class UnsignedIntHandlerTest {
+/**
+ * MySQL data type handler.
+ */
+public interface MySQLDataTypeHandler extends StatelessTypedSPI {
     
-    private final UnsignedIntHandler handler = new UnsignedIntHandler();
-    
-    @Test
-    public void assertHandle() {
-        Serializable actual = handler.handle(1);
-        assertThat(actual, is(1L));
-        actual = handler.handle(-1);
-        assertThat(actual, is(4294967295L));
-    }
+    /**
+     * Handle column value.
+     *
+     * @param value column value
+     * @return handled column value
+     */
+    Serializable handle(Serializable value);
 }
