+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-dialect/shardingsphere-data-pipeline-mysql/src/test/java/org/apache/shardingsphere/data/pipeline/mysql/ingest/column/value/MySQLUnsignedMediumintHandlerTest.java
@@ -17,20 +17,23 @@
 
 package org.apache.shardingsphere.data.pipeline.mysql.ingest.column.value;
 
+import org.apache.shardingsphere.data.pipeline.mysql.ingest.column.value.impl.MySQLUnsignedMediumintHandler;
+import org.junit.Test;
+
 import java.io.Serializable;
 
-public final class UnsignedIntHandler implements ValueHandler {
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertThat;
+
+public final class MySQLUnsignedMediumintHandlerTest {
     
-    private static final long INT_MODULO = 4294967296L;
+    private final MySQLUnsignedMediumintHandler handler = new MySQLUnsignedMediumintHandler();
     
-    @Override
-    public String getTypeName() {
-        return "INT UNSIGNED";
-    }
-    
-    @Override
-    public Serializable handle(final Serializable value) {
-        int intValue = (int) value;
-        return 0 > intValue ? INT_MODULO + intValue : intValue;
+    @Test
+    public void assertHandle() {
+        Serializable actual = handler.handle(1);
+        assertThat(actual, is(1));
+        actual = handler.handle(-1);
+        assertThat(actual, is(16777215));
     }
 }
